+++ blueocean-pipeline-api-impl/src/main/java/io/jenkins/blueocean/preload/BlueUrlTokenizer.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package io.jenkins.blueocean.commons;
+package io.jenkins.blueocean.preload;
 
 import hudson.model.Run;
 import org.kohsuke.accmod.Restricted;
