+++ blueocean-pipeline-api-impl/src/main/java/io/jenkins/blueocean/preload/PipelineBranchRunStatePreloader.java
@@ -27,13 +27,10 @@ import hudson.Extension;
 import hudson.model.Item;
 import hudson.model.Job;
 import hudson.model.Run;
-import io.jenkins.blueocean.commons.BlueUrlTokenizer;
-import io.jenkins.blueocean.commons.RESTFetchPreloader;
 import io.jenkins.blueocean.commons.stapler.Export;
 import io.jenkins.blueocean.rest.factory.BluePipelineFactory;
 import io.jenkins.blueocean.rest.factory.BlueRunFactory;
 import io.jenkins.blueocean.rest.model.BlueRun;
-import io.jenkins.blueocean.service.embedded.rest.AbstractRunImpl;
 import jenkins.branch.MultiBranchProject;
 import jenkins.model.Jenkins;
