+++ blueocean-pipeline-api-impl/src/main/java/io/jenkins/blueocean/preload/PipelineStatePreloader.java
@@ -25,8 +25,6 @@ package io.jenkins.blueocean.preload;
 
 import hudson.Extension;
 import hudson.model.Item;
-import io.jenkins.blueocean.commons.BlueUrlTokenizer;
-import io.jenkins.blueocean.commons.RESTFetchPreloader;
 import io.jenkins.blueocean.commons.stapler.Export;
 import io.jenkins.blueocean.rest.factory.BluePipelineFactory;
 import io.jenkins.blueocean.rest.model.BluePipeline;
